feat: add second version

+++ YourBonoPlatform/Bonds/Domain/Model/Commands/UpdateBondCommand.cs
@@ -5,7 +5,7 @@ public record UpdateBondCommand(
     string Name,
     double NominalValue,
     double MarketValue,
-    double Duration,
+    int Duration,
     int Frequency,
     int InterestRateTypeId,
     double InterestRate,
